ci: Switch shellcheck action to also check formatting
Change the GitHub Actions shellcheck action to use
`luizm/action-sh-checker` as that also checks the formatting with
`shfmt`. Set it up so issues are written to the pull request, although I
may turn this off if it also includes the shfmt diff since I think it is
useless to put that diff on the PR - formatting is not something you fix
manually, you just run `shfmt` and be done with it.