getgo: Restructure to be more robust and cleaner
Restructure getgo to be more robust and to add debugging vars for dryrun
and printing debug info. This is the same as what `getzig` is currently
doing and makes it easier to diagnose when the script goes wrong.

It also checks dependencies first so it does not give strange errors
when deps are not met.