getgo: Robustify getgo some more
Make `getgo` more robust against unknown or missing values returned from
the Go download server. While here, rename the `unstable` arg to
`prerelease` since that is what it really is, and it makes it clearer
why an "unstable" release may not be available. One might always expect
there to be unstable releases, but would only expect prereleases close
to a new release.